Move has_required_fields into separate helpers file
